% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rough_bar.R, R/rough_donut.R, R/rough_line.R,
%   R/rough_scatter.R, R/rough_stackedbar.R
\name{rough_bar}
\alias{rough_bar}
\alias{rough_barh}
\alias{rough_donut}
\alias{rough_pie}
\alias{rough_line}
\alias{rough_scatter}
\alias{rough_stackedbar}
\title{rough_bar}
\usage{
rough_bar(
  data = NULL,
  labels = NULL,
  values = NULL,
  axisFontSize = "1rem",
  axisRoughness = 0.5,
  axisStrokeWidth = 0.5,
  bowing = 0,
  color = "skyblue",
  fillStyle = "hachure",
  fillWeight = 0.5,
  font = "Gaegu",
  highlight = "coral",
  innerStrokeWidth = 1,
  interactive = TRUE,
  labelFontSize = "1rem",
  legend = "",
  padding = 0.1,
  roughness = 1,
  simplification = 0.2,
  stroke = "black",
  strokeWidth = 1,
  title = NULL,
  titleFontSize = "1rem",
  tooltipFontSize = "0.95rem",
  xLabel = NULL,
  yLabel = NULL,
  type = "Bar",
  width = NULL,
  height = NULL,
  elementId = NULL
)

rough_barh(
  data = NULL,
  labels = NULL,
  values = NULL,
  axisFontSize = "1rem",
  axisRoughness = 0.5,
  axisStrokeWidth = 0.5,
  bowing = 0,
  color = "skyblue",
  fillStyle = "hachure",
  fillWeight = 0.5,
  font = "Gaegu",
  highlight = "coral",
  innerStrokeWidth = 1,
  interactive = TRUE,
  labelFontSize = "1rem",
  legend = "",
  padding = 0.1,
  roughness = 1,
  simplification = 0.2,
  stroke = "black",
  strokeWidth = 1,
  title = NULL,
  titleFontSize = "1rem",
  tooltipFontSize = "0.95rem",
  xLabel = NULL,
  yLabel = NULL
)

rough_donut(
  data = NULL,
  labels = NULL,
  values = NULL,
  bowing = 0,
  color = "skyblue",
  fillStyle = "hachure",
  fillWeight = 0.5,
  font = "Gaegu",
  highlight = "coral",
  innerStrokeWidth = 1,
  interactive = TRUE,
  legend = TRUE,
  legendPosition = "right",
  padding = 0.1,
  roughness = 1,
  simplification = 0.2,
  stroke = "black",
  strokeWidth = 1,
  title = NULL,
  titleFontSize = "1rem",
  tooltipFontSize = "0.95rem",
  type = "Donut",
  width = NULL,
  height = NULL,
  elementId = NULL
)

rough_pie(
  data = NULL,
  labels = NULL,
  values = NULL,
  bowing = 0,
  color = "skyblue",
  fillStyle = "hachure",
  fillWeight = 0.85,
  font = "Gaegu",
  highlight = "coral",
  innerStrokeWidth = 0.75,
  interactive = TRUE,
  legend = TRUE,
  legendPosition = "right",
  padding = 0.1,
  roughness = 1,
  simplification = 0.2,
  stroke = "black",
  strokeWidth = 1,
  title = NULL,
  titleFontSize = "1rem",
  tooltipFontSize = "0.95rem",
  type = "Pie"
)

rough_line(
  data,
  y,
  axisFontSize = "1rem",
  axisRoughness = 0.5,
  axisStrokeWidth = 0.5,
  bowing = 0,
  circle = TRUE,
  circleRadius = 10,
  circleRoughnes = 2,
  color = "blue",
  fillStyle = "hachure",
  fillWeight = 0.5,
  font = "Gaegu",
  interactive = TRUE,
  labelFontSize = "1rem",
  legend = TRUE,
  legendPosition = "right",
  roughness = 1,
  simplification = 0.2,
  stroke = "black",
  strokeWidth = 1,
  title = NULL,
  titleFontSize = "1rem",
  tooltipFontSize = "0.95rem",
  xLabel = NULL,
  yLabel = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)

rough_scatter(
  x,
  y,
  axisFontSize = "1rem",
  axisRoughness = 0.5,
  axisStrokeWidth = 0.5,
  bowing = 0,
  color = "blue",
  colorvar = NULL,
  curbZero = FALSE,
  fillStyle = "hachure",
  fillWeight = 0.5,
  font = "Gaegu",
  highlight = "coral",
  highlightLabel = NULL,
  innerStrokeWidth = 1,
  interactive = TRUE,
  labelFontSize = "1rem",
  radius = 8,
  roughness = 1,
  simplification = 0.2,
  stroke = "black",
  strokeWidth = 1,
  title = NULL,
  titleFontSize = "1rem",
  tooltipFontSize = "0.95rem",
  xLabel = NULL,
  yLabel = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)

rough_stackedbar(
  data,
  labels,
  axisFontSize = "1rem",
  axisRoughness = 0.5,
  axisStrokeWidth = 0.5,
  bowing = 0,
  colors = NULL,
  fillStyle = NULL,
  fillWeight = 0.5,
  font = 0,
  highlight = "coral",
  innerStrokeWidth = 1,
  interactive = TRUE,
  labelFontSize = "1rem",
  padding = 0.1,
  roughness = 1,
  simplification = 0.2,
  stroke = "black",
  strokeWidth = 1,
  title = NULL,
  titleFontSize = "1rem",
  tooltipFontSize = "0.95rem",
  xLabel = NULL,
  yLabel = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{in bar barh donut pie is a categorical variable.in stackedbar is a dataframe class}

\item{labels}{labels}

\item{values}{values}

\item{axisFontSize}{: Font-size for axes' labels. Default: '1rem'.}

\item{axisRoughness}{: Roughness for x & y axes. Default: 0.5.}

\item{axisStrokeWidth}{: Stroke-width for x & y axes. Default: 0.5.}

\item{bowing}{: Chart bowing. Default: 0.}

\item{color}{: Color for each bar. Default: 'skyblue'.}

\item{fillStyle}{: Bar fill-style. Should be one of fillStyles shown above.}

\item{fillWeight}{: Weight of inner paths' color. Default: 0.5.}

\item{highlight}{: Color for each bar on hover. Default: 'coral'.}

\item{innerStrokeWidth}{: Stroke-width for paths inside bars. Default: 1.}

\item{interactive}{: Whether or not chart is interactive. Default: true.}

\item{labelFontSize}{: Font-size for axes' labels. Default: '1rem'.}

\item{legend}{: Whether or not to add legend. Default: true.}

\item{padding}{: Padding between bars. Default: 0.1.}

\item{roughness}{: Roughness level of chart. Default: 1.}

\item{simplification}{: Chart simplification. Default 0.2.}

\item{stroke}{: Color of bars' stroke. Default: black.}

\item{strokeWidth}{: Size of bars' stroke. Default: 1.}

\item{title}{: Chart title. Optional.}

\item{titleFontSize}{: Font-size for chart title. Default: '1rem'.}

\item{tooltipFontSize}{: Font-size for tooltip. Default: '0.95rem'.}

\item{xLabel}{: Label for x-axis.}

\item{yLabel}{: Label for y-axis.}

\item{legendPosition}{: Position of legend. Should be either 'left' or 'right'. Default: 'right'.}

\item{circle}{: Whether or not to add circles to chart. Default: true.}

\item{circleRadius}{: Radius of circles. Default: 10.}

\item{highlightLabel}{: If input data is csv or tsv, this should be a column representing what value to display on hover. Otherwise, (x, y) values will be shown on hover.}

\item{radius}{: Circle radius. Default: 8.}

\item{font:}{Font-family to use. You can use 0 or gaegu to use Gaegu, or 1 or indie flower to use Indie Flower. Or feed it something else. Default: Gaegu.}

\item{circleRoughness}{: Roughness of circles. Default: 2.}

\item{colorVar}{: If input data is csv or tsv, this should be an ordinal column with which to color points by.}

\item{burbZero}{: Whether or not to force (x, y) axes to (0, 0). Default: false.}
}
\description{
rough_bar

rough_barh

rough_donut

rough_pie

rough_line

rough_scatter

rough_stackedbar
}
