% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roughViz.R
\name{roughViz}
\alias{roughViz}
\title{Create a Bar Plot}
\usage{
roughViz(
  data,
  roughness,
  type = c("Bar", "BarH", "Donut", "Line", "Pie", "Scatter", "StackedBar"),
  width = NULL,
  height = NULL,
  elementId = NULL,
  ...
)
}
\arguments{
\item{data}{\code{list} A list of named vectors containing the data to be supplied.}

\item{roughness}{\code{character} More roughness correspond to a sloppier looking graph}

\item{...}{Additional optional arguments passed to roughViz.Bar. see \url{https://github.com/jwilber/roughViz#roughvizbar}}
}
\description{
Create a Bar Plot
}
\examples{
roughViz(
  data = list(
    labels = c("one", "two"),
    values = 1:2
  ),
  roughness = 1,
  type = "Donut",
  color = "red",
  font = "gaegu",
  highlight = "orange"
)
}
