% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roughViz.R
\name{roughViz_bar}
\alias{roughViz_bar}
\title{Create a Bar Plot}
\usage{
roughViz_bar(
  labels,
  values,
  roughness,
  width = NULL,
  height = NULL,
  elementId = NULL,
  ...
)
}
\arguments{
\item{labels}{\code{character} The label associated to each bar}

\item{values}{\code{numeric} The value of associated to each bar}

\item{roughness}{\code{character} More roughness correspond to a sloppier looking graph}

\item{...}{Additional optional arguments passed to roughViz.Bar. see \url{https://github.com/jwilber/roughViz#roughvizbar}}
}
\description{
Create a Bar Plot
}
\examples{
roughViz_bar(
  labels    = c("one", "two"),
  values    = 1:2,
  roughness = 1,
  color     = "red",
  font      = "gaegu",
  highlight = "orange"
)
}
